+++ src/modules/init/index.ts
@@ -4,9 +4,9 @@ import * as enquirer from 'enquirer'
 import { outputJson, readJson } from 'fs-extra'
 import * as moment from 'moment'
 import { join } from 'path'
-import { keys, merge, prop } from 'ramda'
+import { keys, merge, prop, reject, test } from 'ramda'
 
-import { getAccount } from '../../conf'
+import { getAccount, getLogin } from '../../conf'
 import log from '../../logger'
 import { MANIFEST_FILE_NAME } from '../../manifest'
 import { promptConfirm } from '../prompts'
@@ -14,6 +14,12 @@ import { promptConfirm } from '../prompts'
 import * as git from './git'
 
 const { mapSeries } = Bluebird
+
+const VTEXInternalTemplates = [
+  // Only show these templates for VTEX e-mail users.
+  'graphql-example',
+  'service-example',
+]
 
 const templates = {
   'graphql-example': 'graphql-example',
@@ -47,6 +53,14 @@ const descriptions = {
   'masterdata-graphql-guide': 'VTEX IO MasterData GraphQL Guide',
   'support app': 'Example of a support app',
 }
+
+const getTemplates = () =>
+  // Return all templates if user's e-mail is `...@vtex...`.
+  // Otherwise filter the VTEX internal templates.
+  test(/\@vtex\./, getLogin()) ?
+    keys(templates) :
+    reject(x => (VTEXInternalTemplates.indexOf(x) >= 0), keys(templates))
+
 
 const promptName = async (repo: string) => {
   const message = 'The app name should only contain numbers, lowercase letters, underscores and hyphens.'
@@ -94,11 +108,13 @@ const promptDescription = async (repo: string) => {
 
 const promptTemplates = async (): Promise<string> => {
   const cancel = 'Cancel'
+  console.log('The templates are' + JSON.stringify(getTemplates()))
+  console.log('The login is ' + getLogin())
   const chosen = prop<string>('service', await enquirer.prompt({
     name: 'service',
     message: 'Choose where do you want to start from',
     type: 'select',
-    choices: [...keys(templates), cancel],
+    choices: [...getTemplates(), cancel],
   }))
   if (chosen === cancel) {
     log.info('Bye o/')